# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: FinDesk AI Analyst API
servers:
  - url: /api
    description: Base API path
security:
  - bearerAuth: []
  - sessionCookie: []
    csrfToken: []
...
paths:
...
  /governance/principals:
    post:
      operationId: issuePrincipal
      tags: [governance]
      summary: Issue a human, service, or manifest-bound agent principal
      security:
        - sessionCookie: []
          csrfToken: []
          stepUpCredential: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PrincipalIssuanceInput" }
      responses:
        "201":
          description: Principal and signed ACTIVATE decision created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PrincipalIssuanceResponse" }
        "400": { $ref: "#/components/responses/InvalidGovernanceInput" }
        "401": { $ref: "#/components/responses/AuthRequired" }
        "403": { $ref: "#/components/responses/AuthForbidden" }
        "409": { $ref: "#/components/responses/IdempotencyConflict" }
        "503": { $ref: "#/components/responses/AuditUnavailable" }
  /governance/principals/{principalId}/revoke:
    post:
      operationId: revokePrincipal
      tags: [governance]
      summary: Append a signed principal revocation decision
      security:
        - sessionCookie: []
          csrfToken: []
          stepUpCredential: []
      parameters:
        - name: principalId
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/GovernanceRevocationInput" }
      responses:
        "200":
          description: Signed principal revocation decision appended
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PrincipalGovernanceDecision" }
        "400": { $ref: "#/components/responses/InvalidGovernanceInput" }
        "401": { $ref: "#/components/responses/AuthRequired" }
        "403": { $ref: "#/components/responses/AuthForbidden" }
        "404": { description: Principal not found, content: { application/json: { schema: { $ref: "#/components/schemas/ApiError" } } } }
        "409": { $ref: "#/components/responses/IdempotencyConflict" }
        "503": { $ref: "#/components/responses/AuditUnavailable" }
  /governance/credentials:
    post:
      operationId: issueCredential
      tags: [governance]
      summary: Issue a permanent application credential and return its plaintext once
      security:
        - sessionCookie: []
          csrfToken: []
          stepUpCredential: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CredentialIssuanceInput" }
      responses:
        "201":
          description: Permanent credential and signed ACTIVATE decision created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CredentialIssuanceResponse" }
        "400": { $ref: "#/components/responses/InvalidGovernanceInput" }
        "401": { $ref: "#/components/responses/AuthRequired" }
        "403": { $ref: "#/components/responses/AuthForbidden" }
        "404": { description: Principal not found, content: { application/json: { schema: { $ref: "#/components/schemas/ApiError" } } } }
        "409": { $ref: "#/components/responses/IdempotencyConflict" }
        "503": { $ref: "#/components/responses/AuditUnavailable" }
  /governance/credentials/{credentialId}/revoke:
    post:
      operationId: revokeCredential
      tags: [governance]
      summary: Append a signed permanent-credential revocation decision
      security:
        - sessionCookie: []
          csrfToken: []
          stepUpCredential: []
      parameters:
        - name: credentialId
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/GovernanceRevocationInput" }
      responses:
        "200":
          description: Signed credential revocation decision appended
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CredentialGovernanceDecision" }
        "400": { $ref: "#/components/responses/InvalidGovernanceInput" }
        "401": { $ref: "#/components/responses/AuthRequired" }
        "403": { $ref: "#/components/responses/AuthForbidden" }
        "404": { description: Credential not found, content: { application/json: { schema: { $ref: "#/components/schemas/ApiError" } } } }
        "409": { $ref: "#/components/responses/IdempotencyConflict" }
        "503": { $ref: "#/components/responses/AuditUnavailable" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: permanent-application-credential
    sessionCookie:
      type: apiKey
      in: cookie
      name: mie_session
    csrfToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
    stepUpCredential:
      type: apiKey
      in: header
      name: X-MIE-Step-Up-Authorization
  responses:
    AuthRequired:
      description: No valid application credential or browser session
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ApiError" }
    AuthForbidden:
      description: Verified principal lacks the required kind, scope, CSRF binding, or resource permission
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ApiError" }
    AuditUnavailable:
      description: Authentication, request audit, or idempotency persistence is unavailable
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ApiError" }
    IdempotencyConflict:
      description: Idempotency key is missing, in progress, or bound to different input
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ApiError" }
    InvalidGovernanceInput:
      description: Governance request or signed-decision input is invalid
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ApiError" }
  schemas:
    PrincipalIssuanceInput:
      type: object
      additionalProperties: false
      required: [principalKind, subject, displayName, scopes, rationale]
      properties:
        principalKind: { type: string, enum: [human, service, agent] }
        subject: { type: string, minLength: 1 }
        displayName: { type: string, minLength: 1 }
        scopes:
          type: array
          minItems: 1
          uniqueItems: true
          items: { type: string, minLength: 1 }
        servicePrincipalId: { type: string, format: uuid }
        manifestId: { type: string, minLength: 1 }
        manifestVersion: { type: string, minLength: 1 }
        rationale: { type: string, minLength: 1 }
    CredentialIssuanceInput:
      type: object
      additionalProperties: false
      required: [principalId, scopes, rationale]
      properties:
        principalId: { type: string, format: uuid }
        scopes:
          type: array
          minItems: 1
          uniqueItems: true
          items: { type: string, minLength: 1 }
        expiresAt: { type: string, format: date-time }
        rationale: { type: string, minLength: 1 }
    GovernanceRevocationInput:
      type: object
      additionalProperties: false
      required: [expectedRevision, expectedDecisionId, rationale]
      properties:
        expectedRevision: { type: integer, minimum: 1 }
        expectedDecisionId: { type: string, format: uuid }
        rationale: { type: string, minLength: 1 }
    PrincipalDecisionSubject:
      type: object
      additionalProperties: false
      required: [subjectType, principalId, principalSha256]
      properties:
        subjectType: { type: string, enum: [PRINCIPAL] }
        principalId: { type: string, format: uuid }
        principalSha256: { type: string, pattern: "^[0-9a-f]{64}$" }
    CredentialDecisionSubject:
      type: object
      additionalProperties: false
      required: [subjectType, credentialId, credentialSha256]
      properties:
        subjectType: { type: string, enum: [CREDENTIAL] }
        credentialId: { type: string, format: uuid }
        credentialSha256: { type: string, pattern: "^[0-9a-f]{64}$" }
    PrincipalGovernanceDecision:
      type: object
      additionalProperties: false
      required: [decisionType, decisionId, verdict, rationale, subject, revision, supersedesDecisionId, humanPrincipalId, credentialId, requestId, decidedAt, nonce, attestationKeyId, attestationHmacSha256]
      properties:
        decisionType: { type: string, enum: [PRINCIPAL] }
        decisionId: { type: string, format: uuid }
        verdict: { type: string, enum: [ACTIVATE, REVOKE] }
        rationale: { type: string, minLength: 1 }
        subject: { $ref: "#/components/schemas/PrincipalDecisionSubject" }
        revision: { type: integer, minimum: 1 }
        supersedesDecisionId: { type: [string, "null"], format: uuid }
        humanPrincipalId: { type: string, format: uuid }
        credentialId: { type: string, format: uuid }
        requestId: { type: string, minLength: 1 }
        decidedAt: { type: string, description: ISO 8601 timestamp with offset }
        nonce: { type: string, minLength: 1 }
        attestationKeyId: { type: string, minLength: 1 }
        attestationHmacSha256: { type: string, pattern: "^[0-9a-f]{64}$" }
    CredentialGovernanceDecision:
      type: object
      additionalProperties: false
      required: [decisionType, decisionId, verdict, rationale, subject, revision, supersedesDecisionId, humanPrincipalId, credentialId, requestId, decidedAt, nonce, attestationKeyId, attestationHmacSha256]
      properties:
        decisionType: { type: string, enum: [CREDENTIAL] }
        decisionId: { type: string, format: uuid }
        verdict: { type: string, enum: [ACTIVATE, REVOKE] }
        rationale: { type: string, minLength: 1 }
        subject: { $ref: "#/components/schemas/CredentialDecisionSubject" }
        revision: { type: integer, minimum: 1 }
        supersedesDecisionId: { type: [string, "null"], format: uuid }
        humanPrincipalId: { type: string, format: uuid }
        credentialId: { type: string, format: uuid }
        requestId: { type: string, minLength: 1 }
        decidedAt: { type: string, description: ISO 8601 timestamp with offset }
        nonce: { type: string, minLength: 1 }
        attestationKeyId: { type: string, minLength: 1 }
        attestationHmacSha256: { type: string, pattern: "^[0-9a-f]{64}$" }
    PrincipalIssuanceResponse:
      type: object
      additionalProperties: false
      required: [principalId, decision]
      properties:
        principalId: { type: string, format: uuid }
        decision: { $ref: "#/components/schemas/PrincipalGovernanceDecision" }
    CredentialIssuanceResponse:
      type: object
      additionalProperties: false
      required: [credentialId, credentialPrefix, permanentCredential, decision]
      properties:
        credentialId: { type: string, format: uuid }
        credentialPrefix: { type: string }
        permanentCredential:
          type: string
          description: Plaintext permanent application credential. Returned once and never recoverable.
        decision: { $ref: "#/components/schemas/CredentialGovernanceDecision" }
...
    ApiError:
      type: object
      properties:
        error:
          type: string
        code:
          type: string
      required:
        - error
...